refract
This is a huge commit.
had an issue with data not being persistent and sort of a gamble if it will work or not.
To solve this I figured I would have to write more modular code and save the data from my api calls so as not to frequently make new request. though there are many tutorials and ways to achieve this  I felt they did not really work or solve the problem I had for various reasons. the solution that I came up with was to write 2 stateless components the first one gets the geolocation of the user and stores it on Local storage and the other get the weather information and stores it. by using local storage,  I eliminated a lot of bugs and issue I had earlier with  REACT states/ props. this makes building the rest of the app much easier since I data will always be available and easier to handle if it is not. although I got a better understanding of states, props and hooks I still have a lot to learn.
To improve the current code I could write the save function in a separate controller and add more CRUD functions.